Split internal cache document string plugin
